update solver tests
