+++ proteus/tests/solver_tests/test_iterative_methods.py
@@ -49,33 +49,6 @@ def create_petsc_vecs(matrix_A):
     x.createWithArray(np.zeros(matrix_A.getSizes()[0][0]))
     return (b, x)
 
-def initialize_schur_ksp_obj(matrix_A, schur_approx):
-    """
-    Creates a right-hand-side and solution PETSc4Py vector for
-    testing ksp solves.
-
-    Parameters
-    ----------
-    matrix_A: :class:`p4pyPETSc.Mat`
-        Global matrix object.
-    schur_approx: :class:`LS.SchurPrecon`
-
-    Returns
-    -------
-    ksp_obj: :class:`p4pyPETSc.KSP`
-    """
-    ksp_obj = p4pyPETSc.KSP().create()
-    ksp_obj.setOperators(matrix_A,matrix_A)
-    pc = schur_approx.pc
-    ksp_obj.setPC(pc)
-    ksp_obj.setFromOptions()
-    pc.setFromOptions()
-    pc.setOperators(matrix_A,matrix_A)
-    pc.setUp()
-    schur_approx.setUp(ksp_obj)
-    ksp_obj.setUp()
-    ksp_obj.pc.setUp()
-    return ksp_obj
 
 def initialize_asm_ksp_obj(matrix_A):
     """
@@ -142,31 +115,6 @@ def build_amg_index_sets(L_sizes):
     return [isvelocity, isu, isv]
 
 
-@pytest.fixture()
-def initialize_petsc_options(request):
-    """Initializes schur complement petsc options. """
-    petsc_options = p4pyPETSc.Options()
-    petsc_options.setValue('ksp_type','gmres')
-    petsc_options.setValue('ksp_gmres_restart',500)
-    petsc_options.setValue('ksp_atol',1e-20)
-    petsc_options.setValue('ksp_gmres_modifiedgramschmidt','')
-    petsc_options.setValue('pc_fieldsplit_type','schur')
-    petsc_options.setValue('pc_fieldsplit_schur_fact_type','upper')
-    petsc_options.setValue('pc_fieldsplit_schur_precondition','user')
-    petsc_options.setValue('fieldsplit_velocity_ksp_type','preonly')
-    petsc_options.setValue('fieldsplit_velocity_pc_type', 'lu')
-    petsc_options.setValue('fieldsplit_pressure_ksp_type','preonly')
-
-@pytest.fixture()
-def initialize_velocity_block_petsc_options(request):
-    petsc_options = p4pyPETSc.Options()
-    petsc_options.setValue('ksp_type','gmres')
-    petsc_options.setValue('ksp_gmres_restart',100)
-    petsc_options.setValue('ksp_pc_side','right')
-    petsc_options.setValue('ksp_atol',1e-8)
-    petsc_options.setValue('ksp_gmres_modifiedgramschmidt','')
-    petsc_options.setValue('pc_type','hypre')
-    petsc_options.setValue('pc_type_hypre_type','boomeramg')
 
 def load_matrix(mat_file_name):
     """Load a matrix """
@@ -175,169 +123,6 @@ def load_matrix(mat_file_name):
                                'import_modules/'+mat_file_name))
     return A
     
-@pytest.fixture()
-def load_nse_cavity_matrix(request):
-    """Loads a Navier-Stokes matrix drawn from the MPRANS module. """
-    A = LAT.petsc_load_matrix(os.path.join
-                              (os.path.dirname(__file__),
-                               'import_modules/NSE_cavity_matrix.bin'))
-    yield A
-
-@pytest.fixture()
-def load_nse_step_matrix(request):
-    """
-    Loads a Navier-Stokes matrix for the backwards step problem from
-    the MPRANS module.  This matrix is constructed using no-slip
-    boundary conditions, and weakly enforced Dirichlet conditions.
-    """
-    A = LAT.petsc_load_matrix(os.path.join
-                              (os.path.dirname(__file__),
-                               'import_modules/NSE_step_no_slip.bin'))
-    yield A
-
-@pytest.fixture()
-def load_small_step_matrix(request):
-    """
-    Loads a small example of a backwards facing step matrix for
-    testing purposes. (Note: this matrix does not have advection)
-    """
-    A = LAT.petsc_load_matrix(os.path.join
-                              (os.path.dirname(__file__),
-                               'import_modules/saddle_point_small.bin'))
-    yield A
-
-@pytest.fixture()
-def load_medium_step_matrix(request):
-    """
-    Loads a medium sized backwards facing step matrix for studying
-    different AMG preconditioners. (Note: this matrix does not have
-    advection)
-    """
-    A = LAT.petsc_load_matrix(os.path.join
-                              (os.path.dirname(__file__),
-                               'import_modules/saddle_point_matrix.bin'))
-    yield A
-
-@pytest.fixture()
-def load_rans2p_step_newton_1(request):
-    A = LAT.petsc_load_matrix(os.path.join
-                              (os.path.dirname(__file__),
-                               'import_modules/rans2p_step_newton_1.bin'))
-    yield A
-
-@pytest.fixture()
-def load_rans2p_step_newton_5(request):
-    A = LAT.petsc_load_matrix(os.path.join
-                              (os.path.dirname(__file__),
-                               'import_modules/rans2p_step_newton_5.bin'))
-    yield A
-
-@pytest.mark.LinearSolvers
-def test_Schur_Sp_solve_global_null_space(load_nse_cavity_matrix,
-                                          initialize_petsc_options):
-    """Tests a KSP solve using the Sp Schur complement approximation.
-    For this test, the global matrix has a null space because the
-    boundary conditions are pure Dirichlet. """
-    mat_A = load_nse_cavity_matrix
-    b, x = create_petsc_vecs(mat_A)
-    petsc_options = initialize_petsc_options
-
-    solver_info = LS.ModelInfo('interlaced',
-                               3,
-                               bdy_null_space=True)
-    schur_approx = LS.Schur_Sp(L=mat_A,
-                               prefix='',
-                               solver_info=solver_info)
-    ksp_obj = initialize_schur_ksp_obj(mat_A,schur_approx)
-    ksp_obj.solve(b,x)
-
-    assert ksp_obj.converged == True
-    assert ksp_obj.its == 35
-    assert np.allclose(ksp_obj.norm, 0.0007464632)
-    assert ksp_obj.reason == 2
-
-@pytest.mark.LinearSolvers
-def test_Schur_Sp_solve(load_nse_step_matrix,
-                        initialize_petsc_options):
-    """Tests a KSP solve using the Sp Schur complement approximation.
-       For this test, the global matrix does not have a null space."""
-    mat_A = load_nse_step_matrix
-    b, x = create_petsc_vecs(mat_A)
-
-    solver_info = LS.ModelInfo('interlaced', 3)
-    schur_approx = LS.Schur_Sp(mat_A,
-                               '',
-                               solver_info=solver_info)
-    ksp_obj = initialize_schur_ksp_obj(mat_A, schur_approx)
-    ksp_obj.solve(b,x)
-
-    assert ksp_obj.converged == True
-    assert ksp_obj.its == 45
-    assert np.allclose(ksp_obj.norm, 394.7036050627)
-    assert ksp_obj.reason == 2
-    
-@pytest.mark.amg
-def test_amg_basic(load_small_step_matrix,
-                   initialize_velocity_block_petsc_options):
-    mat_A = load_small_step_matrix
-
-    petsc_options = initialize_velocity_block_petsc_options
-    L_sizes = mat_A.getSizes()
-    index_sets = build_amg_index_sets(L_sizes)
-
-    #Initialize ksp object
-    F_ksp = initialize_asm_ksp_obj(mat_A.createSubMatrix(index_sets[0],
-                                                      index_sets[0]))
-    b, x = create_petsc_vecs(mat_A.createSubMatrix(index_sets[0],
-                                                index_sets[0]))   
-    F_ksp.solve(b,x)
-    assert F_ksp.its == 9
-
-@pytest.mark.amg
-def test_amg_iteration_performance(load_medium_step_matrix,
-                                   initialize_velocity_block_petsc_options):
-    mat_A = load_medium_step_matrix
-    petsc_options = initialize_velocity_block_petsc_options
-    L_sizes = mat_A.getSizes()
-    index_sets = build_amg_index_sets(L_sizes)
-
-    F_ksp = initialize_asm_ksp_obj(mat_A.createSubMatrix(index_sets[0],
-                                                      index_sets[0]))
-    b, x = create_petsc_vecs(mat_A.createSubMatrix(index_sets[0],
-                                                index_sets[0]))
-
-    F_ksp.solve(b,x)
-    assert F_ksp.its == 41
-
-@pytest.mark.amg
-def test_amg_step_problem_01(load_rans2p_step_newton_1,
-                             initialize_velocity_block_petsc_options):
-    mat_A = load_rans2p_step_newton_1
-    petsc_options = initialize_velocity_block_petsc_options
-    L_sizes = mat_A.getSizes()
-    index_sets = build_amg_index_sets(L_sizes)
-
-    F_ksp = initialize_asm_ksp_obj(mat_A.createSubMatrix(index_sets[0],
-                                                      index_sets[0]))
-    b, x = create_petsc_vecs(mat_A.createSubMatrix(index_sets[0],
-                                                index_sets[0]))
-    F_ksp.solve(b,x)
-    assert F_ksp.its == 59
-
-@pytest.mark.amg
-def test_amg_step_problem_02(load_rans2p_step_newton_5,
-                             initialize_velocity_block_petsc_options):
-    mat_A = load_rans2p_step_newton_5
-    petsc_options = initialize_velocity_block_petsc_options
-    L_sizes = mat_A.getSizes()
-    index_sets = build_amg_index_sets(L_sizes)
-
-    F_ksp = initialize_asm_ksp_obj(mat_A.createSubMatrix(index_sets[0],
-                                                      index_sets[0]))
-    b, x = create_petsc_vecs(mat_A.createSubMatrix(index_sets[0],
-                                                index_sets[0]))
-    F_ksp.solve(b,x)
-    assert F_ksp.its == 60
 
 class TestSmoothingAlgorithms(proteus.test_utils.TestTools.BasicTest):
 
@@ -345,8 +130,6 @@ class TestSmoothingAlgorithms(proteus.test_utils.TestTools.BasicTest):
         self._scriptdir = os.path.dirname(__file__)
         self.saddle_point_matrix=LAT.petsc_load_matrix(os.path.join(self._scriptdir,
                                                                     'import_modules/saddle_point_small'))
-        # self.saddle_point_matrix = LAT.petsc_load_matrix(os.path.join(self._scriptdir,
-        #                                                               'import_modules/saddle_point_matrix'))
     def test_matrix_splitting_1(self):
         vals_F  =    [3.2, 1.1, 5.4, 6.3, 1., -5.1, 1.2]
         col_idx_F  = [0, 1, 2, 0, 2, 0, 1]
